requirements.txt
asgiref==3.9.1
Django==5.2.5
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-decouple==3.8
